[Refactor/#5] NavGraph 그룹화
